enable printf-style argument checking on OS X which (unfortunately) requires adding explicit casts every time a size_t value is output since the type is 'long unsigned int' on OS X, as opposed to 'unsigned int' on Linux and others

+++ applib/APIServer.cc
@@ -192,8 +192,8 @@ MasterAPIServer::run()
 
         if (cc != sizeof(handshake)) {
             log_err("unexpected %d message from %s:%d "
-                    "(expected %d byte handshake)",
-                    cc, intoa(addr), port, sizeof(handshake));
+                    "(expected %u byte handshake)",
+                    cc, intoa(addr), port, (u_int)sizeof(handshake));
             continue;
         }
